Moves the initialization of the header page into separate file
